SDK-adapters merge refactoring - first cut
